Polyphony and full-loop-reset rules (WIP)

+++ src/model/sequencer.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Copy, Serialize, Deserialize)]
 pub struct SequencerNote {
-    pub tone: i32,
+    pub tone: f32,
     pub amplitude: f32,
     pub sustain: f32,
     pub start_beat: f32
